+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.7] - In progress, release soon...
+## [1.0.8] - In progress...
+
+## [1.0.7] - 2021-04-12
+- Add 'default' value support. [#38](https://github.com/benjamin-allion/json-node-normalizer/issues/38)
+- Full 'normalizer.js' refactoring [#41](https://github.com/benjamin-allion/json-node-normalizer/issues/41)
+- <!> Replace 'normalizePath' function by 'normalizePaths' [#41](https://github.com/benjamin-allion/json-node-normalizer/issues/41)
+- <!> New signature for 'normalizeNode' function [#41](https://github.com/benjamin-allion/json-node-normalizer/issues/41)
+- Update documentation for 'normalizePaths' & default values [#41](https://github.com/benjamin-allion/json-node-normalizer/issues/41)
 
 ## [1.0.6] - 2021-03-31
 - Fix dependencies security vulnerability.